+++ src/linter/ModelsimLinter.ts
@@ -1,14 +1,6 @@
-import {
-    workspace,
-    Range,
-    TextDocument,
-    Diagnostic,
-    DiagnosticSeverity,
-    DiagnosticCollection,
-} from 'vscode';
+import * as vscode from 'vscode';
 import * as child from 'child_process';
 import BaseLinter from './BaseLinter';
-import { Logger } from '../logger';
 
 var isWindows = process.platform === 'win32';
 
@@ -18,9 +10,9 @@ export default class ModelsimLinter extends BaseLinter {
     private modelsimWork: string;
     private runAtFileLocation: boolean;
 
-    constructor(diagnosticCollection: DiagnosticCollection, logger: Logger) {
+    constructor(diagnosticCollection: vscode.DiagnosticCollection, logger: vscode.LogOutputChannel) {
         super('modelsim', diagnosticCollection, logger);
-        workspace.onDidChangeConfiguration(() => {
+        vscode.workspace.onDidChangeConfiguration(() => {
             this.getConfig();
         });
         this.getConfig();
@@ -28,28 +20,28 @@ export default class ModelsimLinter extends BaseLinter {
 
     private getConfig() {
         this.modelsimPath = <string>(
-            workspace
+            vscode.workspace
                 .getConfiguration()
                 .get('verilog.linting.path')
         );
         //get custom arguments
         this.modelsimArgs = <string>(
-            workspace
+            vscode.workspace
                 .getConfiguration()
                 .get('verilog.linting.modelsim.arguments')
         );
         this.modelsimWork = <string>(
-            workspace.getConfiguration().get('verilog.linting.modelsim.work')
+            vscode.workspace.getConfiguration().get('verilog.linting.modelsim.work')
         );
         this.runAtFileLocation = <boolean>(
-            workspace
+            vscode.workspace
                 .getConfiguration()
                 .get('verilog.linting.modelsim.runAtFileLocation')
         );
     }
 
-    protected lint(doc: TextDocument) {
-        this.logger.log('modelsim lint requested');
+    protected lint(doc: vscode.TextDocument) {
+        this.logger.info('modelsim lint requested');
         let docUri: string = doc.uri.fsPath; //path of current doc
         let lastIndex: number =
             isWindows == true
@@ -57,7 +49,7 @@ export default class ModelsimLinter extends BaseLinter {
                 : docUri.lastIndexOf('/');
         let docFolder = docUri.substr(0, lastIndex); //folder of current doc
         let runLocation: string =
-            this.runAtFileLocation == true ? docFolder : workspace.rootPath; //choose correct location to run
+            this.runAtFileLocation == true ? docFolder : vscode.workspace.rootPath; //choose correct location to run
         // no change needed for systemverilog
         let command: string =
             this.modelsimPath +
@@ -71,7 +63,7 @@ export default class ModelsimLinter extends BaseLinter {
             command,
             { cwd: runLocation },
             (_error: Error, stdout: string, _stderr: string) => {
-                let diagnostics: Diagnostic[] = [];
+                let diagnostics: vscode.Diagnostic[] = [];
                 let lines = stdout.split(/\r?\n/g);
 
                 // ^\*\* (((Error)|(Warning))( \(suppressible\))?: )(\([a-z]+-[0-9]+\) )?([^\(]*\(([0-9]+)\): )(\([a-z]+-[0-9]+\) )?((((near|Unknown identifier|Undefined variable):? )?["']([\w:;\.]+)["'][ :.]*)?.*)
@@ -80,27 +72,27 @@ export default class ModelsimLinter extends BaseLinter {
                     '^\\*\\* (((Error)|(Warning))( \\(suppressible\\))?: )(\\([a-z]+-[0-9]+\\) )?([^\\(]*)\\(([0-9]+)\\): (\\([a-z]+-[0-9]+\\) )?((((near|Unknown identifier|Undefined variable):? )?["\']([\\w:;\\.]+)["\'][ :.]*)?.*)';
                 // Parse output lines
                 lines.forEach((line, _) => {
-                    let sev: DiagnosticSeverity;
+                    let sev: vscode.DiagnosticSeverity;
                     if (line.startsWith('**')) {
                         let m = line.match(regexExp);
                         try {
                             if (m[7] != doc.fileName) { return; }
                             switch (m[2]) {
                                 case 'Error':
-                                    sev = DiagnosticSeverity.Error;
+                                    sev = vscode.DiagnosticSeverity.Error;
                                     break;
                                 case 'Warning':
-                                    sev = DiagnosticSeverity.Warning;
+                                    sev = vscode.DiagnosticSeverity.Warning;
                                     break;
                                 default:
-                                    sev = DiagnosticSeverity.Information;
+                                    sev = vscode.DiagnosticSeverity.Information;
                                     break;
                             }
                             let lineNum = parseInt(m[8]) - 1;
                             let msg = m[10];
                             diagnostics.push({
                                 severity: sev,
-                                range: new Range(
+                                range: new vscode.Range(
                                     lineNum,
                                     0,
                                     lineNum,
@@ -113,7 +105,7 @@ export default class ModelsimLinter extends BaseLinter {
                         } catch (e) {
                             diagnostics.push({
                                 severity: sev,
-                                range: new Range(0, 0, 0, Number.MAX_VALUE),
+                                range: new vscode.Range(0, 0, 0, Number.MAX_VALUE),
                                 message: line,
                                 code: 'modelsim',
                                 source: 'modelsim',
@@ -121,9 +113,7 @@ export default class ModelsimLinter extends BaseLinter {
                         }
                     }
                 });
-                this.logger.log(
-                    diagnostics.length + ' errors/warnings returned'
-                );
+                this.logger.info(diagnostics.length + ' errors/warnings returned');
                 this.diagnosticCollection.set(doc.uri, diagnostics);
             }
         );