+++ src/providers/DefinitionProvider.ts
@@ -1,30 +1,19 @@
-import {
-    DefinitionProvider,
-    TextDocument,
-    CancellationToken,
-    Position,
-    ProviderResult,
-    DefinitionLink,
-    Range,
-    Definition,
-    LocationLink,
-} from 'vscode';
+import * as vscode from 'vscode';
 import { BsvInfoProviderManger } from '../BsvProvider';
 import { CtagsManager, Symbol } from '../ctags';
-import { Logger } from '../logger';
 
-export class VerilogDefinitionProvider implements DefinitionProvider {
-    private logger: Logger;
-    constructor(logger: Logger) {
+export class VerilogDefinitionProvider implements vscode.DefinitionProvider {
+    private logger: vscode.LogOutputChannel;
+    constructor(logger: vscode.LogOutputChannel) {
         this.logger = logger;
     }
 
     async provideDefinition(
-        document: TextDocument,
-        position: Position,
-        _token: CancellationToken
-    ): Promise<DefinitionLink[] | undefined> {
-        this.logger.log('Definitions Requested: ' + document.uri);
+        document: vscode.TextDocument,
+        position: vscode.Position,
+        _token: vscode.CancellationToken
+    ): Promise<vscode.DefinitionLink[] | undefined> {
+        this.logger.info('Definitions Requested: ' + document.uri);
         // get word start and end
         let textRange = document.getWordRangeAtPosition(position);
         if (!textRange || textRange.isEmpty) { return undefined; }
@@ -32,7 +21,7 @@ export class VerilogDefinitionProvider implements DefinitionProvider {
         let targetText = document.getText(textRange);
         let symbols: Symbol[] = await CtagsManager.getSymbols(document);
         let matchingSymbols: Symbol[] = [];
-        let definitions: DefinitionLink[] = [];
+        let definitions: vscode.DefinitionLink[] = [];
         // find all matching symbols
         for (let i of symbols) {
             if (i.name === targetText) {
@@ -42,24 +31,24 @@ export class VerilogDefinitionProvider implements DefinitionProvider {
         for (let i of matchingSymbols) {
             definitions.push({
                 targetUri: document.uri,
-                targetRange: new Range(
+                targetRange: new vscode.Range(
                     i.startPosition,
-                    new Position(i.startPosition.line, Number.MAX_VALUE)
+                    new vscode.Position(i.startPosition.line, Number.MAX_VALUE)
                 ),
-                targetSelectionRange: new Range(i.startPosition, i.endPosition),
+                targetSelectionRange: new vscode.Range(i.startPosition, i.endPosition),
             });
         }
-        this.logger.log(definitions.length + ' definitions returned');
+        this.logger.info(definitions.length + ' definitions returned');
         return definitions;
     }
 }
 
-export class BsvDefinitionProvider implements DefinitionProvider {
+export class BsvDefinitionProvider implements vscode.DefinitionProvider {
     provideDefinition(
-        document: TextDocument,
-        position: Position,
-        _token: CancellationToken
-    ): ProviderResult<LocationLink[] | Definition> {
+        document: vscode.TextDocument,
+        position: vscode.Position,
+        _token: vscode.CancellationToken
+    ): vscode.ProviderResult<vscode.LocationLink[] | vscode.Definition> {
         const provider = BsvInfoProviderManger.getInstance().getProvider();
         return provider.provideDefinition(document, position);
     }